Sample: Change root package
