+++ sample/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.nostra13.example.universalimageloader"
+    package="com.nostra13.universalimageloader.sample"
     android:versionCode="39"
     android:versionName="1.9.4" >
