+++ sample/src/com/nostra13/universalimageloader/sample/activity/SimpleImageActivity.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright 2011-2013 Sergey Tarasevich
+ * Copyright 2014 Sergey Tarasevich
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,17 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package com.nostra13.example.universalimageloader.activity;
+package com.nostra13.universalimageloader.sample.activity;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentActivity;
-import com.nostra13.example.universalimageloader.Constants;
-import com.nostra13.example.universalimageloader.R;
-import com.nostra13.example.universalimageloader.fragment.ImageGalleryFragment;
-import com.nostra13.example.universalimageloader.fragment.ImageGridFragment;
-import com.nostra13.example.universalimageloader.fragment.ImageListFragment;
-import com.nostra13.example.universalimageloader.fragment.ImagePagerFragment;
+import com.nostra13.universalimageloader.sample.Constants;
+import com.nostra13.universalimageloader.sample.R;
+import com.nostra13.universalimageloader.sample.fragment.ImageGalleryFragment;
+import com.nostra13.universalimageloader.sample.fragment.ImageGridFragment;
+import com.nostra13.universalimageloader.sample.fragment.ImageListFragment;
+import com.nostra13.universalimageloader.sample.fragment.ImagePagerFragment;
 
 /**
  * @author Sergey Tarasevich (nostra13[at]gmail[dot]com)
