+++ sample/src/com/nostra13/universalimageloader/sample/fragment/BaseFragment.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright 2011-2013 Sergey Tarasevich
+ * Copyright 2011-2014 Sergey Tarasevich
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,29 +13,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package com.nostra13.example.universalimageloader.fragment;
+package com.nostra13.universalimageloader.sample.fragment;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
-import com.nostra13.example.universalimageloader.R;
 import com.nostra13.universalimageloader.core.ImageLoader;
+import com.nostra13.universalimageloader.sample.R;
 
 /**
  * @author Sergey Tarasevich (nostra13[at]gmail[dot]com)
  */
 public abstract class BaseFragment extends Fragment {
-
-	protected ImageLoader imageLoader = ImageLoader.getInstance();
-
 	@Override
 	public void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
 
 		setHasOptionsMenu(true);
-		setRetainInstance(true);
 	}
 
 	public void onCreateOptionsMenu(Menu menu, MenuInflater inflater) {
@@ -46,10 +42,10 @@ public abstract class BaseFragment extends Fragment {
 	public boolean onOptionsItemSelected(MenuItem item) {
 		switch (item.getItemId()) {
 			case R.id.item_clear_memory_cache:
-				imageLoader.clearMemoryCache();
+				ImageLoader.getInstance().clearMemoryCache();
 				return true;
 			case R.id.item_clear_disc_cache:
-				imageLoader.clearDiskCache();
+				ImageLoader.getInstance().clearDiskCache();
 				return true;
 			default:
 				return false;
