+++ sample/src/com/nostra13/universalimageloader/sample/fragment/ImageGridFragment.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright 2011-2013 Sergey Tarasevich
+ * Copyright 2011-2014 Sergey Tarasevich
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package com.nostra13.example.universalimageloader.fragment;
+package com.nostra13.universalimageloader.sample.fragment;
 
 import android.graphics.Bitmap;
 import android.os.Bundle;
@@ -23,28 +23,23 @@ import android.view.ViewGroup;
 import android.widget.AdapterView;
 import android.widget.AdapterView.OnItemClickListener;
 import android.widget.BaseAdapter;
+import android.widget.GridView;
 import android.widget.ImageView;
-import android.widget.ListView;
-import android.widget.TextView;
-import com.nostra13.example.universalimageloader.Constants;
-import com.nostra13.example.universalimageloader.R;
+import android.widget.ProgressBar;
 import com.nostra13.universalimageloader.core.DisplayImageOptions;
 import com.nostra13.universalimageloader.core.ImageLoader;
-import com.nostra13.universalimageloader.core.display.FadeInBitmapDisplayer;
-import com.nostra13.universalimageloader.core.display.RoundedBitmapDisplayer;
-import com.nostra13.universalimageloader.core.listener.ImageLoadingListener;
+import com.nostra13.universalimageloader.core.assist.FailReason;
+import com.nostra13.universalimageloader.core.listener.ImageLoadingProgressListener;
 import com.nostra13.universalimageloader.core.listener.SimpleImageLoadingListener;
-
-import java.util.Collections;
-import java.util.LinkedList;
-import java.util.List;
+import com.nostra13.universalimageloader.sample.Constants;
+import com.nostra13.universalimageloader.sample.R;
 
 /**
  * @author Sergey Tarasevich (nostra13[at]gmail[dot]com)
  */
-public class ImageListFragment extends AbsListViewBaseFragment {
+public class ImageGridFragment extends AbsListViewBaseFragment {
 
-	public static final int INDEX = 0;
+	public static final int INDEX = 1;
 
 	String[] imageUrls = Constants.IMAGES;
 
@@ -61,15 +56,15 @@ public class ImageListFragment extends AbsListViewBaseFragment {
 				.cacheInMemory(true)
 				.cacheOnDisk(true)
 				.considerExifParams(true)
-				.displayer(new RoundedBitmapDisplayer(20))
+				.bitmapConfig(Bitmap.Config.RGB_565)
 				.build();
 	}
 
 	@Override
 	public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-		View rootView = inflater.inflate(R.layout.fr_image_list, container, false);
-		listView = (ListView) rootView.findViewById(android.R.id.list);
-		((ListView) listView).setAdapter(new ImageAdapter());
+		View rootView = inflater.inflate(R.layout.fr_image_grid, container, false);
+		listView = (GridView) rootView.findViewById(R.id.grid);
+		((GridView) listView).setAdapter(new ImageAdapter());
 		listView.setOnItemClickListener(new OnItemClickListener() {
 			@Override
 			public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
@@ -79,21 +74,9 @@ public class ImageListFragment extends AbsListViewBaseFragment {
 		return rootView;
 	}
 
-	@Override
-	public void onDestroy() {
-		super.onDestroy();
-		AnimateFirstDisplayListener.displayedImages.clear();
-	}
-
-	private static class ViewHolder {
-		TextView text;
-		ImageView image;
-	}
-
-	class ImageAdapter extends BaseAdapter {
+	public class ImageAdapter extends BaseAdapter {
 
 		private LayoutInflater inflater;
-		private ImageLoadingListener animateFirstListener = new AnimateFirstDisplayListener();
 
 		ImageAdapter() {
 			inflater = LayoutInflater.from(getActivity());
@@ -106,7 +89,7 @@ public class ImageListFragment extends AbsListViewBaseFragment {
 
 		@Override
 		public Object getItem(int position) {
-			return position;
+			return null;
 		}
 
 		@Override
@@ -115,41 +98,50 @@ public class ImageListFragment extends AbsListViewBaseFragment {
 		}
 
 		@Override
-		public View getView(final int position, View convertView, ViewGroup parent) {
-			View view = convertView;
+		public View getView(int position, View convertView, ViewGroup parent) {
 			final ViewHolder holder;
-			if (convertView == null) {
-				view = inflater.inflate(R.layout.item_list_image, parent, false);
+			View view = convertView;
+			if (view == null) {
+				view = inflater.inflate(R.layout.item_grid_image, parent, false);
 				holder = new ViewHolder();
-				holder.text = (TextView) view.findViewById(R.id.text);
-				holder.image = (ImageView) view.findViewById(R.id.image);
+				assert view != null;
+				holder.imageView = (ImageView) view.findViewById(R.id.image);
+				holder.progressBar = (ProgressBar) view.findViewById(R.id.progress);
 				view.setTag(holder);
 			} else {
 				holder = (ViewHolder) view.getTag();
 			}
 
-			holder.text.setText("Item " + (position + 1));
+			ImageLoader.getInstance()
+					.displayImage(imageUrls[position], holder.imageView, options, new SimpleImageLoadingListener() {
+						@Override
+						public void onLoadingStarted(String imageUri, View view) {
+							holder.progressBar.setProgress(0);
+							holder.progressBar.setVisibility(View.VISIBLE);
+						}
 
-			ImageLoader.getInstance().displayImage(imageUrls[position], holder.image, options, animateFirstListener);
+						@Override
+						public void onLoadingFailed(String imageUri, View view, FailReason failReason) {
+							holder.progressBar.setVisibility(View.GONE);
+						}
+
+						@Override
+						public void onLoadingComplete(String imageUri, View view, Bitmap loadedImage) {
+							holder.progressBar.setVisibility(View.GONE);
+						}
+					}, new ImageLoadingProgressListener() {
+						@Override
+						public void onProgressUpdate(String imageUri, View view, int current, int total) {
+							holder.progressBar.setProgress(Math.round(100.0f * current / total));
+						}
+					});
 
 			return view;
 		}
 	}
 
-	private static class AnimateFirstDisplayListener extends SimpleImageLoadingListener {
-
-		static final List<String> displayedImages = Collections.synchronizedList(new LinkedList<String>());
-
-		@Override
-		public void onLoadingComplete(String imageUri, View view, Bitmap loadedImage) {
-			if (loadedImage != null) {
-				ImageView imageView = (ImageView) view;
-				boolean firstDisplay = !displayedImages.contains(imageUri);
-				if (firstDisplay) {
-					FadeInBitmapDisplayer.animate(imageView, 500);
-					displayedImages.add(imageUri);
-				}
-			}
-		}
+	static class ViewHolder {
+		ImageView imageView;
+		ProgressBar progressBar;
 	}
 }
