+++ sample/src/com/nostra13/universalimageloader/sample/widget/UILWidgetProvider.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package com.nostra13.example.universalimageloader.widget;
+package com.nostra13.universalimageloader.sample.widget;
 
 import android.appwidget.AppWidgetManager;
 import android.appwidget.AppWidgetProvider;
@@ -21,14 +21,14 @@ import android.content.Context;
 import android.graphics.Bitmap;
 import android.view.View;
 import android.widget.RemoteViews;
-import com.nostra13.example.universalimageloader.R;
-import com.nostra13.example.universalimageloader.UILApplication;
 import com.nostra13.universalimageloader.core.DisplayImageOptions;
 import com.nostra13.universalimageloader.core.ImageLoader;
 import com.nostra13.universalimageloader.core.assist.ImageSize;
 import com.nostra13.universalimageloader.core.listener.SimpleImageLoadingListener;
+import com.nostra13.universalimageloader.sample.R;
+import com.nostra13.universalimageloader.sample.UILApplication;
 
-import static com.nostra13.example.universalimageloader.Constants.IMAGES;
+import static com.nostra13.universalimageloader.sample.Constants.IMAGES;
 
 /**
  * Example widget provider
